+++ src/models/UpdateOrderPrescriptionRequestPrescription.ts
@@ -107,7 +107,7 @@ export interface UpdateOrderPrescriptionRequestPrescription {
    * @type {CreateOrderRequestPrescriptionsInnerStructuredSig}
    * @memberof UpdateOrderPrescriptionRequestPrescription
    */
-  structuredSig: CreateOrderRequestPrescriptionsInnerStructuredSig;
+  structuredSig?: CreateOrderRequestPrescriptionsInnerStructuredSig | null;
 }
 
 /**
@@ -123,7 +123,6 @@ export function instanceOfUpdateOrderPrescriptionRequestPrescription(
   if (!("quantity" in value) || value["quantity"] === undefined) return false;
   if (!("quantityUnit" in value) || value["quantityUnit"] === undefined) return false;
   if (!("refills" in value) || value["refills"] === undefined) return false;
-  if (!("structuredSig" in value) || value["structuredSig"] === undefined) return false;
   return true;
 }
 
@@ -153,7 +152,10 @@ export function UpdateOrderPrescriptionRequestPrescriptionFromJSONTyped(
     quantity: CreateOrderRequestPrescriptionsInnerQuantityFromJSON(json["quantity"]),
     quantityUnit: json["quantityUnit"],
     refills: json["refills"],
-    structuredSig: CreateOrderRequestPrescriptionsInnerStructuredSigFromJSON(json["structuredSig"]),
+    structuredSig:
+      json["structuredSig"] == null
+        ? undefined
+        : CreateOrderRequestPrescriptionsInnerStructuredSigFromJSON(json["structuredSig"]),
   };
 }